Put query in /core, create /helpers dir, clean up refs
