+++ src/core/helpers/getTimezoneOffset.js
@@ -1,0 +1,3 @@
+module.exports = function(){
+  return new Date().getTimezoneOffset() * -60;
+};
