+++ src/core/helpers/getUrlMaxLength.js
