+++ src/core/helpers/getXhrObject.js
