+++ src/core/helpers/sendBeaconRequest.js
