+++ src/core/query.js
@@ -1,0 +1,72 @@
+var each = require("./utils/each"),
+    events = require("./events"),
+    extend = require("./utils/extend"),
+    getTimezoneOffset = require("./helpers/getTimezoneOffset"),
+    getQueryString = require("./helpers/getQueryString");
+
+function Query(){
+  this.configure.apply(this, arguments);
+};
+extend(Query.prototype, events);
+
+Query.prototype.configure = function(analysisType, params) {
+  this.analysis = analysisType;
+  this.path = '/queries/' + analysisType;
+
+  // Apply params w/ #set method
+  this.params = this.params || {};
+  this.set(params);
+
+  // Localize timezone if none is set
+  if (this.params.timezone === void 0) {
+    this.params.timezone = _getTimezoneOffset();
+  }
+  return this;
+};
+
+Query.prototype.set = function(attributes) {
+  var self = this;
+  each(attributes, function(v, k){
+    var key = k, value = v;
+    if (k.match(new RegExp("[A-Z]"))) {
+      key = k.replace(/([A-Z])/g, function($1) { return "_"+$1.toLowerCase(); });
+    }
+    self.params[key] = value;
+    if (value instanceof Array) {
+      each(value, function(dv, index){
+        if (dv instanceof Array == false && typeof dv === "object") { //  _type(dv)==="Object"
+          each(dv, function(deepValue, deepKey){
+            if (deepKey.match(new RegExp("[A-Z]"))) {
+              var _deepKey = deepKey.replace(/([A-Z])/g, function($1) { return "_"+$1.toLowerCase(); });
+              delete self.params[key][index][deepKey];
+              self.params[key][index][_deepKey] = deepValue;
+            }
+          });
+        }
+      });
+    }
+  });
+  return self;
+};
+
+Query.prototype.get = function(attribute) {
+  var key = attribute;
+  if (key.match(new RegExp("[A-Z]"))) {
+    key = key.replace(/([A-Z])/g, function($1) { return "_"+$1.toLowerCase(); });
+  }
+  if (this.params) {
+    return this.params[key] || null;
+  }
+};
+
+Query.prototype.addFilter = function(property, operator, value) {
+  this.params.filters = this.params.filters || [];
+  this.params.filters.push({
+    "property_name": property,
+    "operator": operator,
+    "property_value": value
+  });
+  return this;
+};
+
+module.exports = Query;
